feat:added and fixed coverage
